// frontend/frontend/src/types/bowler.ts
export type Bowler = {
    bowlerId: number
    bowlerLastName: string | null
    bowlerFirstName: string | null
    bowlerMiddleInit: string | null
    bowlerAddress: string | null
    bowlerCity: string | null
    bowlerState: string | null
    bowlerZip: string | null
    bowlerPhoneNumber: string | null
    teamId: number | null
    teamName: string | null
}